Change StateSupervisionViolationResponseDecision mapping for US_MO (Recidiviz/recidiviz-data#3488)
Updates the mapping of the StateSupervisionViolationResponseDecision enum so that 'CO', # Court Ordered Detention Sanction is mapped to SHOCK_INCARCERATION and 'A', # Capias, is mapped to the new WARRANT_ISSUED value. Includes the migration and updates to the revocations_matrix_filtered_caseload view to handle these new values for the MO dashboard.

GitOrigin-RevId: b1bb544db6a4f5d4f6d172597725ea4c1da6ec97