[Direct ingest] Gracefully handle when process_job task starts and file no longer exists (Recidiviz/recidiviz-data#3959)
I'm pretty sure this crash started happening because the behavior of the google library changed - before we would always get a Blob object, even if the file did not exist. Now get_blob() returns None if the file does not exist. The download_to_temp_file function is built to return None if the file does not exist and we handle that gracefully elsewhere, so just updated to return None here as well.

Closes Recidiviz/recidiviz-data#3893

GitOrigin-RevId: e40e91ce5ab2546fb0b812297c02ddfb0721417c