[PA ingest] Make list ordering in ingest view queries deterministic (Recidiviz/recidiviz-data#4157)
We have a number of places in our PA ingest queries where we're returning columns with results that contain an array or comma-separated list. If we don't make sure those lists are sorted, then the diffing code will think a column with value "A,B,C" is different than "C,B,A" even though the contents are the same.

This PR updates all places in our codebase where we use STRING_AGG or ARRAY_AGG in an ingest view query to make sure the lists are sorted (this happens to only be PA views, good on @terinpw for doing the right thing already in all US_ID views).

Once this lands and is deployed to staging, I'll delete all ingest view files and operations DB rows that resulted from the PA historical refresh and try the ingest view export again.

Work towards Recidiviz/recidiviz-data#4039

GitOrigin-RevId: 89f360b8d99612fba40697afe9c6b5f8500df636