[PA validation] Increase validation threshold for district office accuracy (Recidiviz/recidiviz-data#4189)
Loosens a (very strict) validation threshold because I think the extra null district offices we're seeing in production are legitimate and should stay for now.

More info:
In Recidiviz/recidiviz-data#3866 I updated the supervision periods query to pull in district office info from the PO history table so we could use the actual district office at a point in time instead of the most recent office we know about. This resulted in slightly more null values on periods between when the person starts a supervision stint and is officially assigned a PO for the first time. The nulls triggered our external validation in production on matching district office.

The reason this validation didn't fail in staging - we we didn't drop supervision periods when we reran with the district office updates that created slightly more null periods. The nulls didn't overwrite the original values we had on the supervision periods, which were correct a lot of the time. This is an argument for always dropping objects when we rerun a file!

If we really want to make district office info more accurate (currently .9% error), we can eventually use OUR historical record of the dbo_Release values to determine the district office.

Validation to help us close out Recidiviz/recidiviz-data#4039

GitOrigin-RevId: 031c46868e2a9dff138ca4179f9a0afa6cf7e572