[Deploy] More safety against deploys with local changes (Recidiviz/recidiviz-data#4236)
As part of last night's deploy, Josh committed several local changes in order to fix issues with the deploy scripts that had been introduced recently. This unfortunately made it so that the v1.15.0 tag isn't actually on the master branch (it looks like we skipped it) and will also create issues with the next deploy when the script goes to find the last deployed tag on the master branch and derive the version from that tag.

I've updated the scripts to protect against this and make the error message about protocol clearer when you hit the dirty status error.

GitOrigin-RevId: bd84b0045bcf3c74e3a93199475adbe0385866b5