[SQL preprocessing] Allow ingest views to reduce query complexity by materializing subqueries into temp tables (Recidiviz/recidiviz-data#4346)
In order to fix the PA incarceration periods query ID instability, I want to pull my MASTER_STATE_IDS_FRAGMENT fragment into that query so I can use the master ID. However, this pushes the query complexity over the allowed limit in BigQuery, so I cannot actually run the query once I do this.

This change adds a flag that allows us to build the ingest view query so it puts the raw table views in temp tables, rather than appending them as subqueries in the WITH clause.

The main tradeoff of this change is that it appears to slow queries down by as much as 5x - we probably don't want to apply this new flag broadly until we find an optimization.

Enables fix for Recidiviz/recidiviz-data#4187

GitOrigin-RevId: 4a44c177d1a2c3669a7a73bf9e28cd1186c0350c