[US_ID] add new supervision period and incarceration period termination reasons (Recidiviz/recidiviz-data#4506)
In Idaho, it turns out that when someone is transferred to "history" (aka no longer under IDOC custody), we can get some fine grained details as to why they were released.

A better understanding at least of when someone is released from probation due to early discharge is necessary (Recidiviz/recidiviz-data#4441).

Previously if someone was transferred to history from a prison, we just marked their release_reason as SENTENCE_SERVED. Similarly if someone was released from supervision, we always marked their termination_reason as DISCHARGE. This PR updates that to be more fine grained when we have those details.

This PR:

* Adds COMMUTED, DISMISSED, and PARDONED to the StateSupervisionTerminationReason enum
* For US_ID, when someone is transferred to history, updates the IncarcerationPeriodReleaseReason to be one of: COMMUTED, DEATH, SENTENCE_SERVED (instead of always being sentence_served)
* For US_ID, when someone is transferred to history, updates the SupervisionPeriodReleaseReason to be one of: COMMUTED, DEATH, DISCHARGE, DISMISSED, PARDONED, EXPIRATION (instead of always being discharged)
* Updates the incarceration period ingest view to now expect the loc_ldesc which is used to get this fine grained release information.
* Updates all views for the COVID report and the po_monthly report that use termination/release reasons to now also account for these new reasons.
Co-authored-by: helperbot-recidiviz <helperbot@recidiviz.com>
GitOrigin-RevId: e7126240b33ccb6888ffda028d9fdbbebfe826fd