Add state_code to all objects in recidiviz/tests/persistence (Recidiviz/recidiviz-data#5080)
As part of our ingest mappings overhaul, I want to be able to add more validations to the attrs entity objects themselves since the parser will be setting fields directly on the entity objects, skipping the IngestInfo layer. Included in that validation will be enforcement that nonnull fields like state_code are actually not null. It turns out we are not particularly disciplined about doing this in our tests.

This PR merely adds nonnull state codes to all test entities in the /ingest and /persistence directories. Still to come: calculate.

Work towards Recidiviz/recidiviz-data#3385

GitOrigin-RevId: c20d439a80d3c2a8159efd479a665df04bd7b9fa