Remove apache-airflow from our Pipfile to unblock other upgrades (Recidiviz/recidiviz-data#5092)
Since apache-airflow is installed on the Cloud Composer environment that we run the Airflow DAGs in, the only reason we need to include it in our Pipfile is a) for purposes of code navigability / exploration and 2) to allow a few unit tests to run.

Unfortunately, the airflow library has very strict dependency requirements that have required us to pin the versions of several other libraries we use. Increasingly, we have found that the value of including apache-airflow in our Pipfile is outweighed by need to upgrade certain libraries, including the google client libraries and the attrs library.

This PR merely removes the apache-airflow dependency and unpins cattrs to allow the attrs library to update.

Work towards Recidiviz/recidiviz-data#4287

GitOrigin-RevId: d67b409d5ec4be6dc7a6d22b73562da2f0ab9a5d